Search tags initial commit

+++ src/main/java/com/gejodigital/necesitocr/repositories/TagRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.gejodigital.necesitocr.entities.Tag;
 
 @Repository
-public interface TagRepository extends CrudRepository<Tag, Long>{
+public interface TagRepository extends CrudRepository<Tag, Integer>{
 	public List<Tag> findByName(String name);
 }
